Changed agents.py to use a given eval fun
Changed AlphaBetaAgent to use a given eval function to evaluate a board state in alpha_beta. This will be needed when we're using agents with different evaluation functions to play against each other. Also removed the comment about the game parameter to the constructor since we aren't using that anymore.